* 'develop' of https://github.com/Hanghae99-Team-Splash/Team-Splash-Backend:
  feat/photo-board 기능 구현 완료 및 검수

+++ src/main/java/com/splash/teamsplashbackend/service/PhotoBoardService.java
@@ -1,7 +1,9 @@
 package com.splash.teamsplashbackend.service;
 
 import com.splash.teamsplashbackend.dto.photoBoard.PhotoBoardRequestDto;
+import com.splash.teamsplashbackend.dto.photoBoard.PhotoBoardResponseDto;
 import com.splash.teamsplashbackend.model.PhotoBoard;
+import com.splash.teamsplashbackend.model.User;
 import com.splash.teamsplashbackend.repository.PhotoBoardRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Sort;
@@ -9,17 +11,20 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 public class PhotoBoardService {
     private final PhotoBoardRepository photoBoardRepository;
+//    private final S3Uploader s3Uploader;
 
+//    private final String imageDirName = "static";
 
     public Long uploadPhotoPost(
-//            UserDetailsImpl userDetails,
             PhotoBoardRequestDto photoBoardRequestDto,
-            MultipartFile multipartFile
+            MultipartFile multipartFile,
+            User user
     ) {
         if(multipartFile.getSize() == 0) {
             throw new NullPointerException("등록하려는 게시글에 이미지가 없습니다.");
@@ -27,49 +32,55 @@ public class PhotoBoardService {
 
 //        String imageUrl = s3Uploader.upload(multipartFile, imageDirName);
         PhotoBoard post = PhotoBoard.builder()
-                        .img("imageUrl")
+                        .img("imageUrl 수정 필요")
                         .location(photoBoardRequestDto.getLocation())
                         .description(photoBoardRequestDto.getLocation())
                         .tagname(photoBoardRequestDto.getTagname())
-//                .user()
+                        .user(user)
                         .build();
         photoBoardRepository.save(post);
 
         return post.getId();
     }
 
+
+
     public void editPhotoBoard(
             Long boardId,
-//            User user,
-            PhotoBoardRequestDto photoBoardRequestDto
+            PhotoBoardRequestDto photoBoardRequestDto,
+            User user
     ) {
         PhotoBoard modifiedBoard = photoBoardRepository.findById(boardId).orElseThrow(
                 () -> new NullPointerException("게시글이 없습니다.")
         );
 
-//        if (!modifiedBoard.getUser().getId().equals(user.getId()))
-//            throw new IllegalArgumentException("작성자가 아니라 게시글을 수정 할 수 없습니다.");
-
+        if (!modifiedBoard.getUser().getId().equals(user.getId()))
+            throw new IllegalArgumentException("작성자가 아니라 게시글을 수정 할 수 없습니다.");
 
         modifiedBoard.update(photoBoardRequestDto);
 
         photoBoardRepository.save(modifiedBoard);
     }
 
-    public List<PhotoBoard> findAllPaging() {
+    public List<PhotoBoardResponseDto> findAll() {
+
         List<PhotoBoard> board = photoBoardRepository.findAll(Sort.by(Sort.Direction.DESC,"id"));
 
-//        return board.stream()
-//                .map(
-//                        s-> new BoardResponseDto(
-//                                s.getId(),
-//                                s.getNickname(),
-//                                s.getContent(),
-//                                s.getLocation(),
-//                                s.getImage()))
-//                .collect(Collectors.toList()
-//                );
-        return null;
+        return board.stream()
+                .map(
+                        s-> new PhotoBoardResponseDto(
+                                s.getId(),
+                                s.getUser().getId(),
+                                s.getImg(),
+                                s.getLocation(),
+                                s.getTagname(),
+                                s.getDescription(),
+                                s.getModifiedAt(),
+                                s.getViews()
+                        )
+                )
+                .collect(Collectors.toList()
+                );
     }
 
     public PhotoBoard findPhotoBoard(
@@ -82,14 +93,17 @@ public class PhotoBoardService {
         return photoBoard;
     }
 
-    public void deletePhotoBoard(Long boardId) {
+    public void deletePhotoBoard(
+            Long boardId,
+            Long loginUserId
+    ) {
         PhotoBoard board = photoBoardRepository.findById(boardId)
                 .orElseThrow(
                         () -> new NullPointerException("삭제하려는 게시글이 없습니다.")
                 );
-//        if (!board.getUser().getId().equals(loginUserId)) {
-//            throw new IllegalArgumentException("작성자가 아니라 게시글을 삭제 할 수 없습니다.");
-//        }
+        if (!board.getUser().getId().equals(loginUserId)) {
+            throw new IllegalArgumentException("작성자가 아니라 게시글을 삭제 할 수 없습니다.");
+        }
 
         photoBoardRepository.deleteById(boardId);
     }
